+++ src/Console/Task/DbTask.php
@@ -71,7 +71,7 @@ class DbTask extends Task
             }
             $dump[] = $result['Create Table'] .';';
         }
-        if (!file_put_contents(CONFIG .'/schema/schema.sql', implode("\n\n", $dump))) {
+        if (!file_put_contents(CONFIG .'/db/schema.sql', implode("\n\n", $dump))) {
             $this->error('Error saving schema.sql');
         } 
         return true;
@@ -99,9 +99,9 @@ class DbTask extends Task
     {
         list($plugin,$file) = pluginSplit($name);
         if($plugin){
-            return PLUGINS . DS . Inflector::underscore($plugin) ."/config/schema/{$file}.sql";
+            return PLUGINS . DS . Inflector::underscore($plugin) ."/config/db/{$file}.sql";
         }
-        return CONFIG . "/schema/{$file}.sql";
+        return CONFIG . "/db/{$file}.sql";
     }
 
     private function runSQL($statement,$datasource='default'){